Handle errors in establishing websocket connection
