+++ curl/src/main/scala/org/http4s/curl/websocket/CurlWSClient.scala
@@ -19,18 +19,12 @@ package org.http4s.curl.websocket
 import cats.Foldable
 import cats.effect.IO
 import cats.implicits._
-import org.http4s.Uri
 import org.http4s.client.websocket.WSFrame._
 import org.http4s.client.websocket._
-import org.http4s.curl.internal.Utils
 import org.http4s.curl.unsafe.CurlExecutorScheduler
 import org.http4s.curl.unsafe.CurlRuntime
-import org.http4s.curl.unsafe.libcurl
 import org.http4s.curl.unsafe.libcurl_const
 import scodec.bits.ByteVector
-
-import scala.annotation.unused
-import scala.scalanative.unsafe._
 
 private[curl] object CurlWSClient {
 
@@ -47,70 +41,6 @@ private[curl] object CurlWSClient {
       )
     case _ => IO.raiseError(new RuntimeException("Not running on CurlExecutorScheduler"))
   }
-  final private val ws = Uri.Scheme.unsafeFromString("ws")
-  final private val wss = Uri.Scheme.unsafeFromString("wss")
-
-  private def setup(req: WSRequest, verbose: Boolean)(con: Connection) =
-    Utils.newZone.use { implicit zone =>
-      IO {
-        val scheme = req.uri.scheme.getOrElse(ws)
-
-        if (scheme != ws && scheme != wss)
-          throw new IllegalArgumentException(
-            s"Websocket client can't handle ${scheme.value} scheme!"
-          )
-
-        val uri = req.uri.copy(scheme = Some(scheme))
-
-        con.handler.setCustomRequest(toCString(req.method.renderString))
-
-        if (verbose)
-          con.handler.setVerbose(true)
-
-        con.handler.setUrl(toCString(uri.renderString))
-
-        // NOTE there is no need to handle object lifetime here,
-        // as Connection class and curl handler have the same lifetime
-        con.handler.setWriteData(Utils.toPtr(con))
-        con.handler.setWriteFunction(recvCallback(_, _, _, _))
-
-        con.handler.setHeaderData(Utils.toPtr(con))
-        con.handler.setHeaderFunction(headerCallback(_, _, _, _))
-
-        var headers: Ptr[libcurl.curl_slist] = null
-        req.headers
-          .foreach { header =>
-            headers = libcurl.curl_slist_append(headers, toCString(header.toString))
-          }
-
-        con.handler.setHttpHeader(headers)
-
-      }
-    }
-
-  /** libcurl write callback */
-  private def recvCallback(
-      buffer: Ptr[CChar],
-      size: CSize,
-      nmemb: CSize,
-      userdata: Ptr[Byte],
-  ): CSize =
-    Utils
-      .fromPtr[Connection](userdata)
-      .onReceive(buffer, size, nmemb)
-
-  private def headerCallback(
-      @unused buffer: Ptr[CChar],
-      size: CSize,
-      nitems: CSize,
-      userdata: Ptr[Byte],
-  ): CSize = {
-    Utils
-      .fromPtr[Connection](userdata)
-      .onEstablished()
-
-    size * nitems
-  }
 
   def apply(
       ec: CurlExecutorScheduler,
@@ -121,12 +51,9 @@ private[curl] object CurlWSClient {
   ): Option[WSClient[IO]] =
     Option.when(CurlRuntime.isWebsocketAvailable && CurlRuntime.curlVersionNumber >= 0x75700) {
       WSClient(true) { req =>
-        Connection(recvBufferSize, pauseOn, resumeOn, verbose)
-          .evalTap(setup(req, verbose))
-          .flatTap(con => ec.addHandleR(con.handler.curl, con.onTerminated))
+        Connection(req, ec, recvBufferSize, pauseOn, resumeOn, verbose)
           .map(con =>
             new WSConnection[IO] {
-
               override def send(wsf: WSFrame): IO[Unit] = wsf match {
                 case Close(_, _) =>
                   val flags = libcurl_const.CURLWS_CLOSE
